+++ src/main/java/com/duckblade/osrs/toa/module/TombsOfAmascutModule.java
@@ -20,6 +20,9 @@ import com.duckblade.osrs.toa.features.het.solver.HetSolver;
 import com.duckblade.osrs.toa.features.het.solver.HetSolverOverlay;
 import com.duckblade.osrs.toa.features.hporbs.HpOrbManager;
 import com.duckblade.osrs.toa.features.invocationpresets.InvocationPresetsManager;
+import com.duckblade.osrs.toa.features.nexus.DepositBoxFilter;
+import com.duckblade.osrs.toa.features.nexus.DepositBoxFilterOverlay;
+import com.duckblade.osrs.toa.features.nexus.HelpfulSpiritHighlighter;
 import com.duckblade.osrs.toa.features.pointstracker.PartyPointsTracker;
 import com.duckblade.osrs.toa.features.pointstracker.PointsOverlay;
 import com.duckblade.osrs.toa.features.pointstracker.PointsTracker;
